Refactored client show page
