+++ src/pages/clients/routes.tsx
@@ -34,9 +34,6 @@ const RecurringExpenses = lazy(
 );
 const Statement = lazy(() => import('$app/pages/clients/statement/Statement'));
 const Invoices = lazy(() => import('$app/pages/clients/show/pages/Invoices'));
-const Activities = lazy(
-  () => import('$app/pages/clients/show/pages/Activities')
-);
 const Documents = lazy(() => import('$app/pages/clients/show/pages/Documents'));
 
 export const clientRoutes = (
@@ -106,7 +103,6 @@ export const clientRoutes = (
       <Route path="tasks" element={<Tasks />} />
       <Route path="expenses" element={<Expenses />} />
       <Route path="recurring_expenses" element={<RecurringExpenses />} />
-      <Route path="activities" element={<Activities />} />
       <Route path="documents" element={<Documents />} />
     </Route>
     <Route
