+++ src/pages/clients/show/components/Address.tsx
@@ -8,7 +8,6 @@
  * @license https://www.elastic.co/licensing/elastic-license
  */
 
-import { sanitizeHTML } from '$app/common/helpers/html-string';
 import { useResolveCountry } from '$app/common/hooks/useResolveCountry';
 import { Client } from '$app/common/interfaces/client';
 import { InfoCard } from '$app/components/InfoCard';
@@ -46,15 +45,6 @@ export function Address(props: Props) {
                 </p>
 
                 <p>{resolveCountry(client.country_id)?.name}</p>
-
-                <div className="whitespace-normal max-h-56 overflow-y-auto">
-                  <article
-                    className="prose prose-sm"
-                    dangerouslySetInnerHTML={{
-                      __html: sanitizeHTML(client.public_notes),
-                    }}
-                  />
-                </div>
               </>
             }
             className="h-full"
